move parser types to quartoTypes.ts
Moves QuartoNodeType, QuartoSourcePosition, QuartoSourceLocation,
QuartoNode, QuartoCodeBlock, QuartoRawBlock, QuartoFrontmatter, and
QuartoDocument to quartoTypes.ts. The parser re-exports them for
backwards compatibility.
